vinyl: store max_key for runs in level zero of index tree
To correctly split infiniruns (see single mem plan) into ranges,
we must be able to determine the ranges, in which the infinirun
can be splited.
We will use run pages as units of splitting.
For each page we can use its min_key as left
border and min_key of the next page as right border. For each range,
which intersects with this interval, we will update the compact_priority.

But the last page hasn't right border. To provide the right border for the
last page we store the max_key of all range.

                                Infinirun
               +-------------------------------------------+
               |  page 1  |  page 2  |  page 3  |  page N  |  <--- right border, max_key.
               +-------------------------------------------+
max_ds += 0               max_ds += page2 + page3              max_ds += 0
+-----------+-------------+---------------------+-------------+---------+
| range 1   |   range 2   |       range 3       |   range 4   | range 5 |
+-----------+-------------+---------------------+-------------+---------+
            max_ds += page1                     max_ds += pageN

*max_ds = Max Dump Size

Page N uses infinirun.info.max_key as right border.