vinyl: reset max_dump_size in compact_new()
Assume, that the single mem is implemented. According to
the single mem plan, the dump of infinirange can start during
compaction of any range. If we reset the max_dump_size and
compact_priority in the compact_complete, there is the
following case:
 CP = x           CP = y > x               CP = 0 -> lost y.
    |                 |                       |
    +-----------------|-----------------------+
    |                 compact                 |
    +-----------------|-----------------------+
         +------------+
         |    dump    |
         +------------+

So lets reset the compact priority and max dump size in the
compact_new() and restore the saved values in case of compact abort.